demo-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2

# OpenTelemetry API + SDK
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0

# OTLP exporters
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0

# Auto-instrumentation (0.51b0+ removes pkg_resources dependency)
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-logging==0.51b0
opentelemetry-instrumentation-httpx==0.51b0
